Cleanup: Moved AssemblyData and LoadStatus to separate files.
